The new geometry: the geometry has now a totally different layout that allows to allocate an even or odd spinor at time. The layout of the geometry is organized in this way:
|--bulk even--|--3d border even--|--3d buffer even--|--bulk odd--|--3d border odd--|--3d buffer odd--|--2d border even--|--2d border odd--|--2d buffer even--|--2dbuffer odd--|
|------------------------------------------------------------------glattice---------------------------------------------------------------------------------------------------|
|-----------------------glat_even-------------------|---------------glat_odd-------------------------|

In order to accomodate the same indexing among glattice glat_even and glat_odd the new variable shift has been introduced, it keeps in account of the different point of start of glat_odd.
All the tests of the geometry have been implemented and checked.
All the other test codes have been verified, and they fully work.
Fixed many issues with the intel compiler about float/double conversion.
The RHMC on a comparison run gives the correct plaquette.

BEWARE THE HMC SEEMS NOT TO WORK!!!!
BEWARE THE HMC SEEMS NOT TO WORK!!!!
BEWARE THE HMC SEEMS NOT TO WORK!!!!

Not checked also the wilson loop the spectrum and the modenumber routines.
This release is not meant to be used for production until fully verified.
Antonio






git-svn-id: file:///Users/pica/Downloads/GitClone/svnrep/suN/Branches/HiRep-HMC@477 37a69121-b030-0410-a54a-fdca83ffee47

+++ Inverters/cg_mshift_flt.c
@@ -58,7 +58,7 @@ static int cg_mshift_flt_core(short int *sflags, mshift_par *par, spinor_operato
     /* use out[0] as initial guess */
     M(Mk,&out[0]);
     ++cgiter;
-    spinor_field_mul_add_assign_f_flt(Mk,-par->shift[0],&out[0]);
+    spinor_field_mul_add_assign_f_flt(Mk,((float)(-par->shift[0])),&out[0]);
     spinor_field_sub_f_flt(r,in,Mk);
 
   } else { /* initial guess = 0 for multishift */
@@ -82,23 +82,23 @@ static int cg_mshift_flt_core(short int *sflags, mshift_par *par, spinor_operato
     for (i=0; i<(par->n); ++i) {
       if(sflags[i]) {
         z3[i] = oldomega*z1[i]*z2[i]/(omega*gamma*(z1[i]-z2[i])+z1[i]*oldomega*(1.+par->shift[i]*omega));
-        spinor_field_mul_add_assign_f_flt(&out[i],-omega*z3[i]/z2[i],&p[i]);
+        spinor_field_mul_add_assign_f_flt(&out[i],((float)(-omega*z3[i]/z2[i])),&p[i]);
       }
     }
-    spinor_field_mul_add_assign_f_flt(r,omega,Mk);
+    spinor_field_mul_add_assign_f_flt(r,((float)(omega)),Mk);
     lambda=spinor_field_sqnorm_f_flt(r);
     gamma=lambda/delta;
     delta=lambda;
 
-    spinor_field_mul_f_flt(k,gamma,k);
+    spinor_field_mul_f_flt(k,((float)(gamma)),k);
     spinor_field_add_assign_f_flt(k,r);
     notconverged=0; /* assume that all vectors have converged */
     for (i=0; i<(par->n); ++i) {
       /* check convergence of vectors */
       if(delta*z3[i]*z3[i]>par->err2*innorm2) ++notconverged;
       if(sflags[i]){
-        spinor_field_mul_f_flt(&p[i],gamma*z3[i]*z3[i]/(z2[i]*z2[i]),&p[i]);
-        spinor_field_mul_add_assign_f_flt(&p[i],z3[i],r);
+        spinor_field_mul_f_flt(&p[i],((float)(gamma*z3[i]*z3[i]/(z2[i]*z2[i]))),&p[i]);
+        spinor_field_mul_add_assign_f_flt(&p[i],((float)(z3[i])),r);
         z1[i]=z2[i];
         z2[i]=z3[i];
       }
